chore: move `resolveCircularRef` module to `processOpenAPI` directory
