+++ src/lib/processOpenAPI/resolveCircularRef.ts
